+++ src/Application/Adapter/NotificationPublisherInterface.php
@@ -12,9 +12,11 @@ interface NotificationPublisherInterface
 {
     public function publishHappeningNotification(Happening $happening, string $type, array $data): void;
 
-    public function publishChatMessageNotification(ChatMessageLinkableInterface $object, ChatMessage $message): void;
+    public function publishChatMessageNotification(ChatMessageLinkableInterface $object, ChatMessage $message, int $messageCount): void;
 
     public function publishChatVoteNotification(ChatMessageLinkableInterface $object, ChatMessage $chatMessage, array $votes): void;
 
     public function publishUserConnectionNotification(Sheet $sheet, User $user): void;
+
+    public function publishRequestVisioNotification(Sheet $sheet, User $fromUser, int $toUserId, string $type): void;
 }