+++ src/Domain/Repository/ChatSessionRepositoryInterface.php
@@ -10,7 +10,7 @@ interface ChatSessionRepositoryInterface
 {
     public function add(ChatSession $chatSession): ChatSession;
 
-    public function update(): void;
+    public function update():void;
 
     public function findOneByEventAndUsers(Event $event, User $aUser, User $anotherUser): ?ChatSession;
 
@@ -22,4 +22,6 @@ interface ChatSessionRepositoryInterface
     public function findIdsByEventAndUser(Event $event, User $user): array;
 
     public function findOneById(int $id): ?ChatSession;
+
+    public function hasMessageFromUser (ChatSession $chatSession, User $user): bool;
 }