+++ src/Infrastructure/Bundle/InfrastructureBundle/Migrations/Version20200922131934.php
@@ -1,0 +1,26 @@
+<?php declare(strict_types=1);
+
+namespace Proximum\Vimeet\Infrastructure\Bundle\InfrastructureBundle\Migrations;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+/**
+ * Add call visio dates into event configuration
+ */
+final class Version20200922131934 extends AbstractMigration
+{
+    public function up(Schema $schema) : void
+    {
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE event ADD configuration_call_visio_open_date DATETIME DEFAULT NULL, ADD configuration_call_visio_close_date DATETIME DEFAULT NULL');
+    }
+
+    public function down(Schema $schema) : void
+    {
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE event DROP configuration_call_visio_open_date, DROP configuration_call_visio_close_date');
+    }
+}