+++ src/Ui/Bundle/EventBundle/Resources/assets/js/components/_Chat.js
@@ -9,8 +9,9 @@ function Chat(element) {
     this.addChatFormAction = this.addChatForm.getAttribute('action');
     this.addChatFormSubmit = this.addChatForm.querySelector('button[type="submit"]');
     this.addChatFormList = this.chatContainer.querySelector('.chat-list');
-
+    this.externalListener = [];
     this.chatLoaded = false;
+    this.toUserId = null;
 
     this.addChatForm.addEventListener('submit', this.submitChat.bind(this));
 
@@ -26,6 +27,7 @@ function Chat(element) {
     this.chatVoteDisabledMessage = element.getAttribute('data-chat-vote-disabled-message');
 
     this.chatListeners = [];
+    this.chatMessagesCount = 0;
 }
 
 Chat.prototype.submitChat = function (event) {
@@ -75,10 +77,12 @@ Chat.prototype.initChat = function () {
     $.get(href, function (response) {
         this.removeChatListeners();
         $addChatFormList.empty();
+        let chatMessagesCount = 0;
         response.forEach((item) => {
             const rowEl = document.createElement('div');
             rowEl.id = `chat-message-${item.id}`;
             rowEl.classList.add('chat-row');
+            chatMessagesCount++;
 
             const contentEl = rowEl.appendChild(document.createElement('div'));
             contentEl.classList.add('chat-content');
@@ -202,8 +206,10 @@ Chat.prototype.initChat = function () {
             $addChatFormList[0].appendChild(rowEl);
         });
 
+        this.chatMessagesCount = chatMessagesCount;
         this.addChatFormList.scrollTop = this.addChatFormList.scrollHeight;
         this.chatLoaded = true;
+        this.externalListener.forEach((callback) => callback(response));
     }.bind(this))
         .fail(function (error) {
             console.error('Failed to load chat', error);
@@ -226,6 +232,22 @@ Chat.prototype.reload = function () {
 Chat.prototype.removeChatListeners = function () {
     this.chatListeners.forEach((item) => item[0].removeEventListener('click', item[1]));
     this.chatListeners = [];
+}
+
+Chat.prototype.getChatMessagesCount = function () {
+    return this.chatMessagesCount;
+}
+
+Chat.prototype.addListener = function (callback) {
+    this.externalListener.push(callback);
+}
+
+Chat.prototype.setToUserId = function(toUserId) {
+    this.toUserId = toUserId;
+}
+
+Chat.prototype.getToUserId = function() {
+    return this.toUserId;
 }
 
 export default Chat;