+++ src/Ui/Bundle/EventBundle/Resources/assets/js/main.js
@@ -34,9 +34,11 @@ import CatalogSelectFromNomenclaturesField from './components/_CatalogSelectFrom
 import SortParticipants from './components/_SortParticipants';
 import DateTimePicker from '../../../../../../../assets/js/components/DateTimePicker';
 import addSubmitEventListenerOnElementChange from './components/form/_AddSubmitEventListenerOnElementChange';
+import { showSpinnerOnSubmit } from './components/form/_ShowSpinnerOnSubmit';
 import SheetVideo from './components/_SheetVideo';
 import UserConnectionRegister from './components/_UserConnectionRegister';
 import NotificationToast from './components/_NotificationToast';
+import NotificationCallVisio from './components/_NotificationCallVisio';
 import initNetworking from './networking'
 
 import 'bootstrap';
@@ -274,6 +276,8 @@ function init(target) {
     addSubmitEventListenerOnElementChange(target, 'evaluation', 'evaluation');
     addSubmitEventListenerOnElementChange(target, 'evaluate_meeting', 'evaluation');
 
+    showSpinnerOnSubmit();
+
     if (target.querySelector('[data-agenda-autorefresh]')) {
         new AgendaRefresh();
     }
@@ -285,7 +289,9 @@ function init(target) {
     const userConnectionRegister = new UserConnectionRegister();
     userConnectionRegister.connect();
 
-    initNetworking(document, userConnectionRegister);
+    const notificationCallVisio = new NotificationCallVisio(target.querySelector('[data-notification-call-visio]'), userConnectionRegister);
+
+    initNetworking(document, userConnectionRegister, notificationCallVisio);
 
     new NotificationToast(target.querySelector('[data-notification-toast]'), userConnectionRegister);